models/requirements.txt
category_encoders==2.6.0
Flask==2.2.2
lightgbm==3.3.5
nltk==3.8.1
numpy==1.23.3
pandas==1.5.0
psycopg2_binary==2.9.5
pyspark==3.3.1
scikit_learn==1.2.1
scipy==1.9.3
SQLAlchemy==2.0.0
tqdm==4.64.1
